requirements.txt
matplotlib>=3.3.4
numpy>=1.26.1
sympy>=1.8
setuptools>=50.3.2
z3-solver>=4.8.12.0
cvxpy>=1.1.15
osqp>=0.6.2
Mosek
pandas>=1.2.4
pyyaml
ruamel-yaml>=0.16.12
commonroad-io>=2024.1
# commonroad-vehicle-models>=2.0.0
commonroad-route-planner~=2024.2
# commonroad-dc 
commonroad-drivability-checker~=2024.2
commonroad-crime~=0.4.0

# qp planner
gurobipy
stlpy>=0.3.0

